% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{data}
\name{Dictionary}
\alias{Dictionary}
\title{S4 binding to Dictionary}
\format{\preformatted{Formal class 'jclassName' [package "rJava"] with 2 slots
  ..@ name: chr "org.snu.ids.ha.dic.Dictionary"
  ..@ jobj:Formal class 'jobjRef' [package "rJava"] with 2 slots
  .. .. ..@ jobj  :<externalptr> 
  .. .. ..@ jclass: chr "java/lang/Class"
}}
\usage{
Dictionary
}
\description{
S4 binding to Dictionary
}
\keyword{datasets}

