% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{data}
\name{MorphemeAnalyzer}
\alias{MorphemeAnalyzer}
\title{S4 binding to MorphemeAnalyzer}
\format{\preformatted{Formal class 'jclassName' [package "rJava"] with 2 slots
  ..@ name: chr "org.snu.ids.ha.ma.MorphemeAnalyzer"
  ..@ jobj:Formal class 'jobjRef' [package "rJava"] with 2 slots
  .. .. ..@ jobj  :<externalptr> 
  .. .. ..@ jclass: chr "java/lang/Class"
}}
\usage{
MorphemeAnalyzer
}
\description{
S4 binding to MorphemeAnalyzer
}
\keyword{datasets}

