% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{data}
\name{Parser}
\alias{Parser}
\title{S4 binding to Parser}
\format{\preformatted{Formal class 'jclassName' [package "rJava"] with 2 slots
  ..@ name: chr "org.snu.ids.ha.sp.Parser"
  ..@ jobj:Formal class 'jobjRef' [package "rJava"] with 2 slots
  .. .. ..@ jobj  :<externalptr> 
  .. .. ..@ jclass: chr "java/lang/Class"
}}
\usage{
Parser
}
\description{
S4 binding to Parser
}
\keyword{datasets}

