% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{data}
\name{RMorphemeAnalyzer}
\alias{RMorphemeAnalyzer}
\title{S4 binding to RMorphemeAnalyzer}
\format{\preformatted{Formal class 'jclassName' [package "rJava"] with 2 slots
  ..@ name: chr "io.github.youhyunjo.rkma.RMorphemeAnalyzer"
  ..@ jobj:Formal class 'jobjRef' [package "rJava"] with 2 slots
  .. .. ..@ jobj  :<externalptr> 
  .. .. ..@ jclass: chr "java/lang/Class"
}}
\usage{
RMorphemeAnalyzer
}
\description{
S4 binding to RMorphemeAnalyzer
}
\keyword{datasets}

